+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- [Core] Fix classNames injected by `<IconButton>` will be overridden with custom `className`.
+
 ## [1.4.0]
 ### Added
 - [Form] Add transition to text in `<TextInputRow>` when being focused.